+++ src/core/utils/common.tsx
@@ -47,13 +47,35 @@ export const getUserColor = (emailOrName: string) => {
 /**
  * Get the container as a child of body, or create one if it doesn't exist.
  * Mostly used for dynamic components like Modal or Toast.
+ *
+ * @param containerId id of the container to find or create
+ * @param ownerDocument Can be changed if the container should be in a different document (e.g. in popup).
  */
-export const getContainer = (containerId: string) => {
-  let container = document.getElementById(containerId);
-  if (container == null) {
-    container = document.createElement('div');
+export const getContainer = (
+  containerId: string,
+  ownerDocument = getDocument(),
+) => {
+  let container = ownerDocument?.getElementById(containerId) ?? undefined;
+  if (container == null && !!ownerDocument) {
+    container = ownerDocument.createElement('div');
     container.setAttribute('id', containerId);
-    document.body.appendChild(container);
+    ownerDocument.body.appendChild(container);
   }
   return container;
+};
+
+/**
+ * Get document if it is defined.
+ * Used to support SSR/SSG applications.
+ */
+export const getDocument = () => {
+  return typeof document === 'undefined' ? undefined : document;
+};
+
+/**
+ * Get window if it is defined.
+ * Used to support SSR/SSG applications.
+ */
+export const getWindow = () => {
+  return typeof window === 'undefined' ? undefined : window;
 };