+++ stories/core/DatePicker.stories.tsx
@@ -6,16 +6,18 @@ import { action } from '@storybook/addon-actions';
 import { useEffect, useState } from '@storybook/addons';
 import { Meta, Story } from '@storybook/react';
 import React from 'react';
-import { DatePicker, IconButton } from '../../src/core';
+import { DatePicker, IconButton, TimePicker } from '../../src/core';
 import {
   DatePickerProps,
   generateLocalizedStrings,
 } from '../../src/core/DatePicker/DatePicker';
 import SvgCalendar from '@itwin/itwinui-icons-react/cjs/icons/Calendar';
+import { CreeveyMeta } from 'creevey';
 
 export default {
   title: 'Core/DatePicker',
   component: DatePicker,
+  subcomponents: { TimePicker },
   argTypes: {
     onChange: { control: { disable: true } },
     className: { control: { disable: true } },
@@ -23,50 +25,31 @@ export default {
     date: { control: 'date' },
     setFocus: { defaultValue: true },
   },
-} as Meta<DatePickerProps>;
+  parameters: {
+    creevey: {
+      ignoreElements: ['#picker-button + span'],
+      tests: {
+        async open() {
+          const button = await this.browser.findElement({
+            id: 'picker-button',
+          });
+
+          await this.browser.actions().click(button).perform();
+          await this.expect(await this.takeScreenshot()).to.matchImage(
+            'opened',
+          );
+        },
+      },
+    },
+  },
+} as Meta<DatePickerProps> & CreeveyMeta;
 
 export const Basic: Story<DatePickerProps> = (args) => {
-  const { date = new Date(), setFocus = true, localizedNames } = args;
-  const [opened, setOpened] = useState(false);
-  const [currentDate, setCurrentDate] = useState(new Date(date));
-  const onChange = (date: Date) => {
-    setCurrentDate(date);
-    action(`New date value: ${date}`, { clearOnStoryChange: false })();
-  };
-
-  useEffect(() => {
-    setCurrentDate(new Date(date));
-    return () => action('', { clearOnStoryChange: true })();
-  }, [date]);
-  return (
-    <>
-      <IconButton onClick={() => setOpened(!opened)}>
-        <SvgCalendar />
-      </IconButton>
-      <span style={{ marginLeft: 16 }}>{currentDate.toString()}</span>
-      {opened && (
-        <div style={{ marginTop: 4 }}>
-          <DatePicker
-            date={currentDate}
-            onChange={onChange}
-            localizedNames={localizedNames}
-            setFocus={setFocus}
-          />
-        </div>
-      )}
-    </>
-  );
-};
-
-Basic.args = {
-  date: new Date(),
-};
-
-export const Localized: Story<DatePickerProps> = (args) => {
   const {
-    date = new Date(),
+    date = new Date(2021, 4, 11, 14, 55, 22),
     setFocus = true,
-    localizedNames = generateLocalizedStrings('ja'),
+    localizedNames,
+    ...rest
   } = args;
   const [opened, setOpened] = useState(false);
   const [currentDate, setCurrentDate] = useState(new Date(date));
@@ -81,13 +64,104 @@ export const Localized: Story<DatePickerProps> = (args) => {
   }, [date]);
   return (
     <>
-      <IconButton onClick={() => setOpened(!opened)}>
+      <IconButton onClick={() => setOpened(!opened)} id='picker-button'>
         <SvgCalendar />
       </IconButton>
       <span style={{ marginLeft: 16 }}>{currentDate.toString()}</span>
       {opened && (
         <div style={{ marginTop: 4 }}>
           <DatePicker
+            {...rest}
+            date={currentDate}
+            onChange={onChange}
+            localizedNames={localizedNames}
+            setFocus={setFocus}
+          />
+        </div>
+      )}
+    </>
+  );
+};
+
+Basic.args = {
+  date: new Date(2021, 4, 11, 14, 55, 22),
+};
+
+export const WithTime: Story<DatePickerProps> = (args) => {
+  const {
+    date = new Date(2021, 4, 11, 14, 55, 22),
+    setFocus = true,
+    showTime = true,
+    localizedNames,
+    ...rest
+  } = args;
+  const [opened, setOpened] = useState(false);
+  const [currentDate, setCurrentDate] = useState(new Date(date));
+  const onChange = (date: Date) => {
+    setCurrentDate(date);
+    action(`New date value: ${date}`, { clearOnStoryChange: false })();
+  };
+
+  useEffect(() => {
+    setCurrentDate(new Date(date));
+    return () => action('', { clearOnStoryChange: true })();
+  }, [date]);
+  return (
+    <>
+      <IconButton onClick={() => setOpened(!opened)} id='picker-button'>
+        <SvgCalendar />
+      </IconButton>
+      <span style={{ marginLeft: 16 }}>{currentDate.toString()}</span>
+      {opened && (
+        <div style={{ marginTop: 4 }}>
+          <DatePicker
+            {...rest}
+            date={currentDate}
+            onChange={onChange}
+            localizedNames={localizedNames}
+            setFocus={setFocus}
+            showTime={showTime}
+          />
+        </div>
+      )}
+    </>
+  );
+};
+
+WithTime.args = {
+  date: new Date(2021, 4, 11, 14, 55, 22),
+  setFocus: true,
+  showTime: true,
+};
+
+export const Localized: Story<DatePickerProps> = (args) => {
+  const {
+    date = new Date(2021, 4, 11, 14, 55, 22),
+    setFocus = true,
+    localizedNames = generateLocalizedStrings('ja'),
+    ...rest
+  } = args;
+  const [opened, setOpened] = useState(false);
+  const [currentDate, setCurrentDate] = useState(new Date(date));
+  const onChange = (date: Date) => {
+    setCurrentDate(date);
+    action(`New date value: ${date}`, { clearOnStoryChange: false })();
+  };
+
+  useEffect(() => {
+    setCurrentDate(new Date(date));
+    return () => action('', { clearOnStoryChange: true })();
+  }, [date]);
+  return (
+    <>
+      <IconButton onClick={() => setOpened(!opened)} id='picker-button'>
+        <SvgCalendar />
+      </IconButton>
+      <span style={{ marginLeft: 16 }}>{currentDate.toString()}</span>
+      {opened && (
+        <div style={{ marginTop: 4 }}>
+          <DatePicker
+            {...rest}
             date={currentDate}
             onChange={onChange}
             localizedNames={localizedNames}
@@ -100,6 +174,6 @@ export const Localized: Story<DatePickerProps> = (args) => {
 };
 
 Localized.args = {
-  date: new Date(),
+  date: new Date(2021, 4, 11, 14, 55, 22),
   localizedNames: generateLocalizedStrings('ja'),
 };